% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econdataverse-package.R
\docType{package}
\name{econdataverse-package}
\alias{econdataverse}
\alias{econdataverse-package}
\title{econdataverse: Easily Load and Install the 'EconDataverse'}
\description{
The 'EconDataverse' is a universe of open-source packages to work seamlessly with economic data. This package is designed to make it easy to install and load multiple 'EconDataverse' packages in a single step. Learn more about the 'EconDataverse' at \url{https://www.econdataverse.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidy-intelligence/r-econdataverse}
  \item \url{https://tidy-intelligence.github.io/r-econdataverse/}
  \item Report bugs at \url{https://github.com/tidy-intelligence/r-econdataverse/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph@tidy-intelligence.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID}) [copyright holder]

}
\keyword{internal}
