% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{econdataverse_deps}
\alias{econdataverse_deps}
\title{List all econdataverse dependencies}
\usage{
econdataverse_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
econdataverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
List all econdataverse dependencies
}
