% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{econdataverse_packages}
\alias{econdataverse_packages}
\title{List all packages in the econdataverse}
\usage{
econdataverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include econdataverse in the list?}
}
\value{
A character vector of package names.
}
\description{
List all packages in the econdataverse
}
\examples{
econdataverse_packages()
}
