% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{econdataverse_update}
\alias{econdataverse_update}
\title{Update econdataverse packages}
\usage{
econdataverse_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
econdataverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
This will check to see if all econdataverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
econdataverse_update()
}
}
